create a separate component for book rendering
